+++ app/Http/Controllers/VolunteeringController.php
@@ -4,19 +4,30 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\VolunteeringRequest;
 use App\Models\Volunteering;
+use App\Models\Category;
+use App\Models\Accessibility;
+use App\Models\VolunteeringSuitability;
 use Illuminate\Http\Request;
+use Inertia\Inertia;
+use Illuminate\Support\Facades\Gate;
+
+use Redirect;
 
 class VolunteeringController extends Controller
 {
+
     /**
      * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function index(Request $request)
     {
-        $volunteering = Volunteering::all();
+        $this->authorize('viewAny', Volunteering::class);
 
-        return view('volunteering.index', compact('volunteering'));
+        return Inertia::render('Volunteering/Index', [
+            'volunteerings' => \Auth::user()->currentTeam->volunteerings()->with('subteam:id,name')->get()
+        ]);;
+
     }
 
     /**
@@ -25,7 +36,16 @@ class VolunteeringController extends Controller
      */
     public function create(Request $request)
     {
-        return view('volunteering.create');
+        $this->authorize('create', Volunteering::class);
+
+        return Inertia::render('Volunteering/Form', [
+            'categories' => Category::select('id','title')->get(),
+            'accessibilities' => Accessibility::select('id','title')->get(),
+            'suitabilities' => VolunteeringSuitability::select('id','title')->get(),
+            'requirements' => config('system.requirements'),
+            'subteams' => \Auth::user()->currentTeam->subteams()->select('id','name')->get(),
+            'team' => \Auth::user()->currentTeam()->select('name','phone','email')->first()
+        ]);
     }
 
     /**
@@ -34,56 +54,99 @@ class VolunteeringController extends Controller
      */
     public function store(VolunteeringRequest $request)
     {
-        $volunteering = Volunteering::create($request->validated());
+        $this->authorize('create', Volunteering::class);
 
-        $request->session()->flash('volunteering.id', $volunteering->id);
+        $volunteering = Volunteering::create(array_merge($request->validated(), ['created_by' => \Auth::user()->id]));
+        $volunteering->categories()->sync($request->categories);
+        $volunteering->accessibilities()->sync($request->accessibilities);
+        $volunteering->suitabilities()->sync($request->suitabilities);
 
-        return redirect()->route('volunteering.index');
+        return Redirect::route('volunteering.edit', [
+            'volunteering' => $volunteering
+        ]);
+
     }
 
     /**
      * @param \Illuminate\Http\Request $request
-     * @param \App\Models\Volunteering $volunteeringOpportunity
+     * @param \App\Models\Volunteering $volunteering
      * @return \Illuminate\Http\Response
      */
     public function show(Request $request, Volunteering $volunteering)
     {
-        return view('volunteering.show', compact('volunteering'));
+        // return Inertia::render('Volunteering/Form', [
+        //     'volunteering' => $volunteering
+        // ]);
     }
 
     /**
      * @param \Illuminate\Http\Request $request
-     * @param \App\Models\Volunteering $volunteeringOpportunity
+     * @param \App\Models\Volunteering $volunteering
      * @return \Illuminate\Http\Response
      */
-    public function edit(Request $request, Volunteering $volunteering)
+    public function edit(Volunteering $volunteering)
     {
-        return view('volunteering.edit', compact('volunteering'));
+        $this->authorize('update', $volunteering);
+
+        $volunteering->categories = $volunteering->categories()->allRelatedIds()->toArray();
+        $volunteering->accessibilities = $volunteering->accessibilities()->allRelatedIds()->toArray();
+        $volunteering->suitabilities = $volunteering->suitabilities()->allRelatedIds()->toArray();
+
+        return Inertia::render('Volunteering/Form', [
+            'volunteering' => $volunteering,
+            'categories' => Category::select('id','title')->get(),
+            'accessibilities' => Accessibility::select('id','title')->get(),
+            'suitabilities' => VolunteeringSuitability::select('id','title')->get(),
+            'requirements' => config('system.requirements'),
+            'skills' => config('system.skills'),
+            'subteams' => \Auth::user()->currentTeam->subteams()->select('id','name')->get()
+        ]);
     }
 
     /**
      * @param \App\Http\Requests\VolunteeringRequest $request
-     * @param \App\Models\Volunteering $volunteeringOpportunity
+     * @param \App\Models\Volunteering $volunteering
      * @return \Illuminate\Http\Response
      */
     public function update(VolunteeringRequest $request, Volunteering $volunteering)
     {
+        $this->authorize('update', $volunteering);
+
         $volunteering->update($request->validated());
+        $volunteering->categories()->sync($request->categories);
+        $volunteering->accessibilities()->sync($request->accessibilities);
+        $volunteering->suitabilities()->sync($request->suitabilitiesx);
+        $volunteering->update(['updated_by' => \Auth::user()->id]);
 
-        $request->session()->flash('volunteering.id', $volunteering->id);
-
-        return redirect()->route('volunteering.index');
+        return Redirect::route('volunteering.edit', compact('volunteering'));
     }
 
     /**
      * @param \Illuminate\Http\Request $request
-     * @param \App\Models\Volunteering $volunteeringOpportunity
+     * @param \App\Models\Volunteering $volunteering
      * @return \Illuminate\Http\Response
      */
     public function destroy(Request $request, Volunteering $volunteering)
     {
-        $volunteering->delete();
+        $this->authorize('delete', $volunteering);
 
-        return redirect()->route('volunteering.index');
+        $volunteering->delete();
+        return Redirect::route('volunteerings.show');
+    }
+
+    public function destroyAll(Request $request, $volunteering_ids)
+    {
+        $volunteering_ids_array = explode('-', $volunteering_ids);
+        $volunteerings_query = Volunteering::whereIn('id', $volunteering_ids_array);
+
+        $volunteerings = $volunteerings_query->get();
+
+        foreach($volunteerings as $volunteering) {
+            $this->authorize('delete', $volunteering);
+        }
+
+        $volunteerings_query->delete();
+
+        return Redirect::route('volunteerings.show');
     }
 }
